Refactoring - Isolation login service from user service
