+++ src/main/java/com/hwangwongyu/news/login/SessionKey.java
@@ -1,0 +1,7 @@
+package com.hwangwongyu.news.login;
+
+public class SessionKey {
+
+    public static final String USER_LOGIN_INFO = "userLoginInfo";
+
+}
