+++ src/main/java/com/hwangwongyu/news/user/UserService.java
@@ -1,8 +1,5 @@
 package com.hwangwongyu.news.user;
 
-import com.hwangwongyu.news.redis.UserLoginInfo;
-
-import javax.servlet.http.HttpSession;
 import java.util.List;
 
 public interface UserService {
@@ -20,10 +17,6 @@ public interface UserService {
     UserDTO findUserByLoginId(String loginId);
 
     UserDTO findUserByNickname(String nickname);
-
-    UserDTO loginUser(UserLoginInfo userLoginInfo);
-
-    void logout(HttpSession httpSession);
 
     Boolean sendEmailAuthNCode(String toEmail);
